Progress

+++ src/base/building.controller.ts
@@ -2,9 +2,11 @@ import {
   Body,
   ClassSerializerInterceptor,
   Controller,
+  ForbiddenException,
   Param,
   ParseIntPipe,
   Post,
+  Req,
   UseGuards,
   UseInterceptors,
 } from '@nestjs/common';
@@ -20,6 +22,7 @@ import {
 import { DequeueBuildingRequestDto } from '../empire/request/dequeue-building-request.dto';
 import { DowngradeBuildingRequestDto } from '../empire/request/downgrade-building-request.dto';
 import { AuthGuard } from '@nestjs/passport';
+import { BaseService } from './base.service';
 
 @ApiTags('Building')
 @Controller('base')
@@ -27,14 +30,21 @@ import { AuthGuard } from '@nestjs/passport';
 @ApiBearerAuth()
 @UseInterceptors(ClassSerializerInterceptor)
 export class BuildingController {
-  constructor(private readonly buildingService: BuildingService) {}
+  constructor(
+    private readonly buildingService: BuildingService,
+    private readonly baseService: BaseService,
+  ) {}
   @Post(':id/building/build')
   @ApiOkResponse()
   @ApiConflictResponse()
   async buildBuilding(
     @Param('id', ParseIntPipe) id: number,
     @Body() data: BuildBuildingRequestDto,
+    @Req() req: any,
   ) {
+    if (!(await this.baseService.isOwner(id, req.user.empireId))) {
+      throw new ForbiddenException();
+    }
     await this.buildingService.buildBuilding(id, data.building);
   }
 
@@ -44,7 +54,11 @@ export class BuildingController {
   async queueBuilding(
     @Param('id', ParseIntPipe) id: number,
     @Body() data: QueueBuildingRequestDto,
+    @Req() req: any,
   ) {
+    if (!(await this.baseService.isOwner(id, req.user.empireId))) {
+      throw new ForbiddenException();
+    }
     await this.buildingService.queueBuilding(id, data.building);
   }
 
@@ -54,14 +68,21 @@ export class BuildingController {
   async dequeueBuilding(
     @Param('id', ParseIntPipe) id: number,
     @Body() data: DequeueBuildingRequestDto,
+    @Req() req: any,
   ) {
+    if (!(await this.baseService.isOwner(id, req.user.empireId))) {
+      throw new ForbiddenException();
+    }
     await this.buildingService.unqueueBuilding(id, data.index);
   }
 
   @Post(':id/building/cancel')
   @ApiOkResponse()
   @ApiConflictResponse()
-  async cancelBuilding(@Param('id', ParseIntPipe) id: number) {
+  async cancelBuilding(@Param('id', ParseIntPipe) id: number, @Req() req: any) {
+    if (!(await this.baseService.isOwner(id, req.user.empireId))) {
+      throw new ForbiddenException();
+    }
     await this.buildingService.cancelBuilding(id);
   }
 
@@ -71,9 +92,11 @@ export class BuildingController {
   async downgradeBuilding(
     @Param('id', ParseIntPipe) id: number,
     @Body() data: DowngradeBuildingRequestDto,
+    @Req() req: any,
   ) {
+    if (!(await this.baseService.isOwner(id, req.user.empireId))) {
+      throw new ForbiddenException();
+    }
     await this.buildingService.deleteBuilding(id, data.building);
   }
-
-  // -- RESEARCH --
 }
